+++ intrastat_base/views/intrastat.xml
@@ -17,7 +17,7 @@
               parent="account.menu_finance_configuration" sequence="50"/>
 
 
-    <record id="intrastat_result_view" model="ir.ui.view">
+    <record id="intrastat_result_view_form" model="ir.ui.view">
       <field name="name">intrastat.result_view_form</field>
       <field name="model">intrastat.result.view</field>
       <field name="arch" type="xml">
