+++ intrastat_base/views/product_template.xml
