+++ intrastat_base/views/res_company.xml
