+++ intrastat_base/views/res_partner.xml
