+++ intrastat_product/__init__.py
@@ -1,7 +1,3 @@
 # -*- encoding: utf-8 -*-
 
-from . import res_company
-from . import intrastat
-from . import intrastat_declaration
-from . import stock
-from . import account_invoice
+from . import models
