--- intrastat_product/intrastat_view.xml
@@ -1,203 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-    Copyright (C) 2010-2015 Akretion (http://www.akretion.com/)
-    Copyright (C) 2015 Noviat (http://www.noviat.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    @author Luc De Meyer <luc.demeyer@noviat.com>
-    The licence is in the file __openerp__.py
--->
-
-<openerp>
-  <data>
-
-    <!-- Add the H.S. code menu under the Intrastat Config menu -->
-    <menuitem id="intrastat_code_menu"
-        action="product_harmonized_system.hs_code_action"
-        parent="intrastat_base.menu_intrastat_config_root"
-        sequence="10"/>
-
-    <!-- Inherit tree view of H.S. code -->
-    <record id="hs_code_tree" model="ir.ui.view">
-      <field name="name">intrastat.hs.code.tree</field>
-      <field name="model">hs.code</field>
-      <field name="inherit_id" ref="product_harmonized_system.hs_code_tree"/>
-      <field name="arch" type="xml">
-        <field name="local_code" position="after">
-          <field name="intrastat_unit_id"/>
-        </field>
-      </field>
-    </record>
-
-    <!-- Inherit form view for H.S. code -->
-    <record id="hs_code_form" model="ir.ui.view">
-      <field name="name">intrastat.hs.code.form</field>
-      <field name="model">hs.code</field>
-      <field name="inherit_id" ref="product_harmonized_system.hs_code_form"/>
-      <field name="arch" type="xml">
-        <field name="local_code" position="after">
-          <field name="intrastat_unit_id"/>
-        </field>
-      </field>
-    </record>
-
-    <!-- Intrastat Supplementary Unit -->
-   <record id="intrastat_unit_form" model="ir.ui.view">
-      <field name="name">intrastat.unit.form</field>
-      <field name="model">intrastat.unit</field>
-      <field name="arch" type="xml">
-        <form string="Intrastat Supplementary Unit">
-          <group>
-            <field name="name"/>
-            <field name="uom_id" required="1"/>
-            <field name="description"/>
-            <field name="active"/>
-          </group>
-        </form>
-      </field>
-    </record>
-
-    <record id="intrastat_unit_tree" model="ir.ui.view">
-      <field name="name">intrastat.unit.tree</field>
-      <field name="model">intrastat.unit</field>
-      <field name="arch" type="xml">
-        <tree string="Intrastat Supplementary Units">
-          <field name="name"/>
-          <field name="uom_id"/>
-          <field name="description"/>
-        </tree>
-      </field>
-    </record>
-
-    <record id="intrastat_unit_search" model="ir.ui.view">
-      <field name="name">intrastat.unit.search</field>
-      <field name="model">intrastat.unit</field>
-      <field name="arch" type="xml">
-        <search string="Search Intrastat Supplementary Units">
-          <field name="name"
-              filter_domain="['|', ('name', 'ilike', self), ('description', 'ilike', self)]"/>
-            <group string="Group By" name="groupby">
-                <filter name="uom_groupby" string="Regular UoM"
-                    context="{'group_by': 'uom_id'}"/>
-            </group>
-        </search>
-      </field>
-    </record>
-
-
-    <record id="intrastat_unit_action" model="ir.actions.act_window">
-      <field name="name">Supplementary Units</field>
-      <field name="res_model">intrastat.unit</field>
-      <field name="view_mode">tree,form</field>
-    </record>
-
-    <menuitem id="intrastat_unit_menu"
-        action="intrastat_unit_action"
-        parent="intrastat_base.menu_intrastat_config_root"
-        sequence="40"/>
-
-    <!-- Intrastat Transaction Type -->
-    <record id="intrastat_transaction_form" model="ir.ui.view">
-      <field name="name">intrastat.transaction_form</field>
-      <field name="model">intrastat.transaction</field>
-      <field name="arch" type="xml">
-        <form string="Intrastat Transaction Type">
-          <group>
-            <field name="code"/>
-            <field name="description"/>
-            <field name="company_id" groups="base.group_multi_company"/>
-          </group>
-        </form>
-      </field>
-    </record>
-
-    <record id="intrastat_transaction_tree" model="ir.ui.view">
-      <field name="name">intrastat.transaction_tree</field>
-      <field name="model">intrastat.transaction</field>
-      <field name="arch" type="xml">
-        <tree string="Intrastat Transaction Types">
-          <field name="code"/>
-          <field name="description"/>
-          <field name="company_id" groups="base.group_multi_company"/>
-        </tree>
-      </field>
-    </record>
-
-    <record id="intrastat_transaction_mode_search" model="ir.ui.view">
-      <field name="name">intrastat.transaction.mode.search</field>
-      <field name="model">intrastat.transaction</field>
-      <field name="arch" type="xml">
-        <search string="Search Intrastat Transaction Types">
-          <field name="description" string="Code or Description"
-            filter_domain="['|', ('code', 'ilike', self), ('description', 'ilike', self)]"/>
-            <group string="Group By" name="groupby">
-                <filter name="company_groupby" string="Company"
-                    context="{'group_by': 'company_id'}"/>
-            </group>
-        </search>
-      </field>
-    </record>
-
-    <record id="intrastat_transaction_action" model="ir.actions.act_window">
-      <field name="name">Transaction Types</field>
-      <field name="res_model">intrastat.transaction</field>
-      <field name="view_mode">tree,form</field>
-    </record>
-
-    <menuitem id="intrastat_transaction_menu"
-        action="intrastat_transaction_action"
-        parent="intrastat_base.menu_intrastat_config_root"
-        sequence="20"/>
-
-
-    <!-- Intrastat Transport Mode -->
-    <record id="intrastat_transport_mode_form" model="ir.ui.view">
-      <field name="name">intrastat.transport.mode.form</field>
-      <field name="model">intrastat.transport_mode</field>
-      <field name="arch" type="xml">
-        <form string="Intrastat Transport Mode">
-          <group>
-            <field name="name"/>
-            <field name="code"/>
-            <field name="description"/>
-          </group>
-        </form>
-      </field>
-    </record>
-
-    <record id="intrastat_transport_mode_tree" model="ir.ui.view">
-      <field name="name">intrastat.transport.mode.tree</field>
-      <field name="model">intrastat.transport_mode</field>
-      <field name="arch" type="xml">
-        <tree string="Intrastat Transportat Modes">
-          <field name="code"/>
-          <field name="name"/>
-          <field name="description"/>
-        </tree>
-      </field>
-    </record>
-
-    <record id="intrastat_transport_mode_search" model="ir.ui.view">
-      <field name="name">intrastat.transport.mode.search</field>
-      <field name="model">intrastat.transport_mode</field>
-      <field name="arch" type="xml">
-        <search string="Search Intrastat Transport Modes">
-          <field name="name" string="Name, Code or Description"
-              filter_domain="['|', '|', ('name', 'ilike', self), ('description', 'ilike', self), ('code', 'ilike', self)]"/>
-        </search>
-      </field>
-    </record>
-
-    <record id="intrastat_transport_action" model="ir.actions.act_window">
-      <field name="name">Transport Modes</field>
-      <field name="res_model">intrastat.transport_mode</field>
-      <field name="view_mode">tree,form</field>
-    </record>
-
-    <menuitem id="intrastat_transport_menu"
-        action="intrastat_transport_action"
-        parent="intrastat_base.menu_intrastat_config_root"
-        sequence="30"/>
-
-  </data>
-</openerp>
