+++ intrastat_product/models/__init__.py
@@ -1,0 +1,10 @@
+# -*- encoding: utf-8 -*-
+
+from . import res_company
+from . import hs_code
+from . import intrastat_transaction
+from . import intrastat_unit
+from . import intrastat_transport_mode
+from . import intrastat_product_declaration
+from . import stock_picking
+from . import account_invoice
