+++ intrastat_product/models/intrastat_transport_mode.py
@@ -1,6 +1,7 @@
 # -*- encoding: utf-8 -*-
 ##############################################################################
 #
+#    Intrastat Product module for Odoo
 #    Copyright (C) 2011-2015 Akretion (http://www.akretion.com)
 #    Copyright (C) 2009-2015 Noviat (http://www.noviat.com)
 #    @author Alexis de Lattre <alexis.delattre@akretion.com>
@@ -21,24 +22,28 @@
 #
 ##############################################################################
 
-{
-    'name': 'Product Harmonized System Codes',
-    'version': '0.2',
-    'category': 'Reporting',
-    'license': 'AGPL-3',
-    'summary': 'Base module for Product Import/Export reports',
-    'author': 'Akretion, Noviat, Odoo Community Association (OCA)',
-    'depends': ['product'],
-    'conflicts': ['report_intrastat'],
-    'data': [
-        'security/product_hs_security.xml',
-        'security/ir.model.access.csv',
-        'views/hs_code.xml',
-        'views/product_category.xml',
-        'views/product_template.xml',
-    ],
-    'demo': [
-        'demo/product_demo.xml',
-    ],
-    'installable': True,
-}
+from openerp import models, fields, api
+
+
+class IntrastatTransportMode(models.Model):
+    _name = 'intrastat.transport_mode'
+    _description = "Intrastat Transport Mode"
+    _rec_name = 'display_name'
+    _order = 'code'
+
+    display_name = fields.Char(
+        string='Display Name', compute='_display_name', store=True,
+        readonly=True)
+    code = fields.Char(string='Code', required=True)
+    name = fields.Char(string='Name', required=True, translate=True)
+    description = fields.Char(string='Description', translate=True)
+
+    @api.one
+    @api.depends('name', 'code')
+    def _display_name(self):
+        self.display_name = '%s. %s' % (self.code, self.name)
+
+    _sql_constraints = [(
+        'intrastat_transport_code_unique',
+        'UNIQUE(code)',
+        'Code must be unique.')]
