+++ intrastat_product/models/intrastat_unit.py
@@ -1,6 +1,7 @@
 # -*- encoding: utf-8 -*-
 ##############################################################################
 #
+#    Intrastat Product module for Odoo
 #    Copyright (C) 2011-2015 Akretion (http://www.akretion.com)
 #    Copyright (C) 2009-2015 Noviat (http://www.noviat.com)
 #    @author Alexis de Lattre <alexis.delattre@akretion.com>
@@ -21,24 +22,20 @@
 #
 ##############################################################################
 
-{
-    'name': 'Product Harmonized System Codes',
-    'version': '0.2',
-    'category': 'Reporting',
-    'license': 'AGPL-3',
-    'summary': 'Base module for Product Import/Export reports',
-    'author': 'Akretion, Noviat, Odoo Community Association (OCA)',
-    'depends': ['product'],
-    'conflicts': ['report_intrastat'],
-    'data': [
-        'security/product_hs_security.xml',
-        'security/ir.model.access.csv',
-        'views/hs_code.xml',
-        'views/product_category.xml',
-        'views/product_template.xml',
-    ],
-    'demo': [
-        'demo/product_demo.xml',
-    ],
-    'installable': True,
-}
+from openerp import models, fields
+
+
+class IntrastatUnit(models.Model):
+    _name = 'intrastat.unit'
+    _description = 'Intrastat Supplementary Units'
+
+    name = fields.Char(
+        string='Name', required=True)
+    description = fields.Char(
+        string='Description', required=True)
+    uom_id = fields.Many2one(
+        'product.uom', string='Regular UoM',
+        help="Select the regular Unit of Measure of Odoo that corresponds "
+        "to this Intrastat Supplementary Unit.")
+    active = fields.Boolean(
+        string='Active', default=True)
