+++ intrastat_product/models/res_company.py
@@ -1,10 +1,11 @@
 # -*- encoding: utf-8 -*-
 ##############################################################################
 #
-#    Copyright (c) 2012-2015 Noviat nv/sa (www.noviat.com)
-#    Copyright (C) 2015 Akretion (http://www.akretion.com)
-#    @author Luc de Meyer <info@noviat.com>
+#    Intrastat Product module for Odoo
+#    Copyright (C) 2011-2015 Akretion (http://www.akretion.com)
+#    Copyright (C) 2009-2015 Noviat (http://www.noviat.com)
 #    @author Alexis de Lattre <alexis.delattre@akretion.com>
+#    @author Luc de Meyer <info@noviat.com>
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as
@@ -13,11 +14,11 @@
 #
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 #    GNU Affero General Public License for more details.
 #
 #    You should have received a copy of the GNU Affero General Public License
-#    along with this program. If not, see <http://www.gnu.org/licenses/>.
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
 
@@ -26,6 +27,19 @@ from openerp import models, fields, api
 
 class ResCompany(models.Model):
     _inherit = 'res.company'
+
+    intrastat_arrivals = fields.Selection(
+        '_intrastat_arrivals', string='Arrivals',
+        default='extended', required=True)
+    intrastat_dispatches = fields.Selection(
+        '_intrastat_arrivals', string='Dispatches',
+        default='extended', required=True)
+    intrastat_transport_id = fields.Many2one(
+        'intrastat.transport_mode',
+        string='Default Transport Mode', ondelete='restrict')
+    intrastat = fields.Char(
+        string='Intrastat Declaration', store=True, readonly=True,
+        compute='_compute_intrastat')
 
     @api.model
     def _intrastat_arrivals(self):
@@ -52,16 +66,3 @@ class ResCompany(models.Model):
             self.intrastat = 'extended'
         else:
             self.intrastat = 'standard'
-
-    intrastat_arrivals = fields.Selection(
-        '_intrastat_arrivals', string='Arrivals',
-        default='extended', required=True)
-    intrastat_dispatches = fields.Selection(
-        '_intrastat_arrivals', string='Dispatches',
-        default='extended', required=True)
-    intrastat_transport_id = fields.Many2one(
-        'intrastat.transport_mode',
-        string='Default Transport Mode', ondelete='restrict')
-    intrastat = fields.Char(
-        string='Intrastat Declaration', store=True, readonly=True,
-        compute='_compute_intrastat')
