+++ intrastat_product/views/intrastat_product_declaration.xml
@@ -2,7 +2,7 @@
 <openerp>
   <data>
 
-    <record id="intrastat_product_declaration_form" model="ir.ui.view">
+    <record id="intrastat_product_declaration_view_form" model="ir.ui.view">
       <field name="name">intrastat.product.declaration.form</field>
       <field name="model">intrastat.product.declaration</field>
       <field name="arch" type="xml">
@@ -117,8 +117,8 @@
       <field name="description">Intrastat Product Declaration Validated</field>
     </record>
     
-    <record id="intrastat_product_tree" model="ir.ui.view">
-      <field name="name">intrastat.product.tree</field>
+    <record id="intrastat_product_declaration_view_tree" model="ir.ui.view">
+      <field name="name">intrastat.product.declaration.tree</field>
       <field name="model">intrastat.product.declaration</field>
       <field name="arch" type="xml">
         <tree string="Intrastat Product Declarations" colors="blue:state=='draft'">
@@ -133,8 +133,8 @@
       </field>
     </record>
 
-    <record id="intrastat_product_search" model="ir.ui.view">
-      <field name="name">intrastat.product.search</field>
+    <record id="intrastat_product_declaration_view_search" model="ir.ui.view">
+      <field name="name">intrastat.product.declaration.search</field>
       <field name="model">intrastat.product.declaration</field>
       <field name="arch" type="xml">
         <search string="Search Intrastat Product Declarations">
@@ -150,8 +150,8 @@
       </field>
     </record>
 
-    <record id="intrastat_product_graph" model="ir.ui.view">
-      <field name="name">intrastat.product.graph</field>
+    <record id="intrastat_product_declaration_view_graph" model="ir.ui.view">
+      <field name="name">intrastat.product.declaration.graph</field>
       <field name="model">intrastat.product.declaration</field>
       <field name="arch" type="xml">
         <graph string="Intrastat Product" type="bar">
@@ -162,16 +162,7 @@
       </field>
     </record>
 
-    <record id="intrastat_product_tree_action" model="ir.actions.act_window">
-      <field name="name">Intrastat Product Declaration</field>
-      <field name="res_model">intrastat.product.declaration</field>
-      <field name="view_mode">tree,form,graph</field>
-    </record>
-
-    <!-- TO BE REMOVED -->
-    <menuitem id="intrastat_product_menu"
-      parent="intrastat_base.menu_intrastat_base_root"
-      action="intrastat_product_tree_action"/>
+    <!-- No menuitem nor action since these are provided by the localization modules -->
 
   </data>
 </openerp>
