+++ intrastat_product/views/intrastat_transport_mode.xml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!--
+    Copyright (C) 2010-2015 Akretion (http://www.akretion.com/)
+    Copyright (C) 2015 Noviat (http://www.noviat.com/)
+    @author Alexis de Lattre <alexis.delattre@akretion.com>
+    @author Luc De Meyer <luc.demeyer@noviat.com>
+    The licence is in the file __openerp__.py
+-->
+
+<openerp>
+  <data>
+
+    <!-- Intrastat Transport Mode -->
+    <record id="intrastat_transport_mode_form" model="ir.ui.view">
+      <field name="name">intrastat.transport.mode.form</field>
+      <field name="model">intrastat.transport_mode</field>
+      <field name="arch" type="xml">
+        <form string="Intrastat Transport Mode">
+          <group>
+            <field name="name"/>
+            <field name="code"/>
+            <field name="description"/>
+          </group>
+        </form>
+      </field>
+    </record>
+
+    <record id="intrastat_transport_mode_tree" model="ir.ui.view">
+      <field name="name">intrastat.transport.mode.tree</field>
+      <field name="model">intrastat.transport_mode</field>
+      <field name="arch" type="xml">
+        <tree string="Intrastat Transportat Modes">
+          <field name="code"/>
+          <field name="name"/>
+          <field name="description"/>
+        </tree>
+      </field>
+    </record>
+
+    <record id="intrastat_transport_mode_search" model="ir.ui.view">
+      <field name="name">intrastat.transport.mode.search</field>
+      <field name="model">intrastat.transport_mode</field>
+      <field name="arch" type="xml">
+        <search string="Search Intrastat Transport Modes">
+          <field name="name" string="Name, Code or Description"
+              filter_domain="['|', '|', ('name', 'ilike', self), ('description', 'ilike', self), ('code', 'ilike', self)]"/>
+        </search>
+      </field>
+    </record>
+
+    <record id="intrastat_transport_action" model="ir.actions.act_window">
+      <field name="name">Transport Modes</field>
+      <field name="res_model">intrastat.transport_mode</field>
+      <field name="view_mode">tree,form</field>
+    </record>
+
+    <menuitem id="intrastat_transport_menu"
+        action="intrastat_transport_action"
+        parent="intrastat_base.menu_intrastat_config_root"
+        sequence="30"/>
+
+  </data>
+</openerp>
