+++ intrastat_product/views/res_company.xml
