+++ intrastat_product/views/stock_picking.xml
