+++ product_harmonized_system/__init__.py
@@ -1,4 +1,3 @@
 # -*- encoding: utf-8 -*-
 
-from . import hs
-from . import product
+from . import models
