+++ product_harmonized_system/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- encoding: utf-8 -*-
+
+from . import hs_code
+from . import product_category
+from . import product_template
