+++ product_harmonized_system/views/hs_code.xml
@@ -10,7 +10,7 @@
   <data>
 
     <!-- Search view for H.S. code -->
-    <record id="hs_code_search" model="ir.ui.view">
+    <record id="hs_code_view_search" model="ir.ui.view">
       <field name="name">hs.code.search</field>
       <field name="model">hs.code</field>
       <field name="arch" type="xml">
@@ -22,7 +22,7 @@
     </record>
 
     <!-- Tree view for H.S. code -->
-    <record id="hs_code_tree" model="ir.ui.view">
+    <record id="hs_code_view_tree" model="ir.ui.view">
       <field name="name">hs.code.tree</field>
       <field name="model">hs.code</field>
       <field name="arch" type="xml">
@@ -36,7 +36,7 @@
     </record>
 
     <!-- Form view for H.S. code -->
-    <record id="hs_code_form" model="ir.ui.view">
+    <record id="hs_code_view_form" model="ir.ui.view">
       <field name="name">hs.code.form</field>
       <field name="model">hs.code</field>
       <field name="arch" type="xml">
